requirements-test.txt
--find-links wheels

shuup==1.10.14
mock==2.0.0
pytest==3.0.6
pytest-cov==2.4.0
pytest-django==3.1.2
pytest-sugar==0.8.0
flake8==3.3.0
flake8-isort==2.1.3
isort==4.2.5
beautifulsoup4==4.5.3
faker==1.0.0
-e .
